refactor(ui): improve naming in the theme implementation
